add more filters, add more unit tests
