+++ lib/src/builders/filterBuilder.ts
@@ -392,61 +392,3 @@ export class FilterBuilder {
             : { items: this.filters }
     }
 }
-
-export type Pagination = {
-    take: number;
-    skip: number;
-}
-
-export class PaginationBuilder {
-    private pageNumber: number = 1;
-    private pageSize: number = 10;
-
-    public setPageSize(pageSize: number): PaginationBuilder {
-        if (pageSize < 0) {
-            throw new Error("pageSize can not be below 0")
-        }
-
-        this.pageSize = pageSize;
-
-        return this;
-    }
-
-    /**
-     * Page size is index 1 based
-     * @param pageNumber 
-     * @returns 
-     */
-    public setPage(pageNumber: number): PaginationBuilder {
-        if (pageNumber < 1) {
-            throw new Error("pageNumber can not be below 1")
-        }
-
-        this.pageNumber = pageNumber;
-
-        return this;
-    }
-
-    build(): Pagination {
-        return {
-            take: this.pageSize,
-            skip: (this.pageNumber - 1) * this.pageSize
-        }
-    }
-}
-
-export class RelevanceModifierBuilder {
-    private modifiers: RelevanceModifier[] = [];
-
-    public reset(): this {
-        this.modifiers = [];
-
-        return this;
-    }
-
-    public build(): RelevanceModifierCollection | null {
-        return this.modifiers.length === 0
-            ? null
-            : { items: this.modifiers }
-    }
-}
