+++ lib/src/builders/index.ts
@@ -1,3 +1,4 @@
 export * from "./settings";
 export * from "./filterBuilder";
+export * from "./paginationBuilder";
 export * from "./search";
